# server/controllers/jobs.js
const Job = require("../models/Job");
const {StatusCodes} = require("http-status-codes");
const {BadRequestError, NotFoundError} = require("../errors");

const get10Jobs = async (req, res) => {
  const page = req.query.page;
  const count = await Job.find({createdBy: req.user.userId}).countDocuments();
  const jobs = await Job.find({createdBy: req.user.userId})
    .sort("-createdAt")
    .skip(page * 10 - 10)
    .limit(10);
  res.status(StatusCodes.OK).json({jobs, count: count});
};

const getJob = async (req, res) => {
  const {
    user: {userId},
    params: {id: jobId},
  } = req;
  const job = await Job.findOne({
    _id: jobId,
    createdBy: userId,
  });
  if (!job) throw new NotFoundError(`No job found with ID no. ${jobId}.`);
  res.status(StatusCodes.OK).json({job});
};

const createJob = async (req, res) => {
  req.body.createdBy = req.user.userId;
  const job = await Job.create(req.body);
  res.status(StatusCodes.CREATED).json({job});
};

const updateJob = async (req, res) => {
  const {
    body: {company, position},
    user: {userId},
    params: {id: jobId},
  } = req;
  if (company === "" || position === "")
    throw new BadRequestError("Company or position fields cannot be empty.");
  const job = await Job.findOneAndUpdate(
    {_id: jobId, createdBy: userId},
    req.body,
    {new: true, runValidators: true}
  );
  if (!job) throw new NotFoundError(`No job found with ID no. ${jobId}.`);
  res.status(StatusCodes.OK).json({job});
};

const deleteJob = async (req, res) => {
  const {
    user: {userId},
    params: {id: jobId},
  } = req;
  const job = await Job.findOneAndRemove({
    _id: jobId,
    createdBy: userId,
  });
  if (!job) throw new NotFoundError(`No job found with ID no. ${jobId}.`);
  res
    .status(StatusCodes.OK)
    .json({msg: `Job ID no. ${jobId} has been deleted.`});
};

module.exports = {
  get10Jobs,
  getJob,
  createJob,
  updateJob,
  deleteJob,
};